requirements.txt
annotated-types==0.7.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
docopt==0.6.2
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.0.0
googleapis-common-protos==1.67.0rc1
httplib2==0.22.0
idna==3.10
mysql-connector-python==9.2.0
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.0a1
pydantic_core==2.28.0
pyparsing==3.2.1
pyzmq==26.2.1
requests==2.32.3
rsa==4.9
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
yarg==0.1.10
